requirements.txt
aiofiles
aiohttp
asyncio
dnspython
ffmpeg-python
gitpython
hachoir
unidecode
heroku3
motor==3.0.0
pillow
psutil
py-tgcalls==0.8.6
pykeyboard
pyrogram==1.4.16
python-dotenv
pyyaml
requests
speedtest-cli
spotipy
tgcrypto
numpy
git+https://github.com/yt-dlp/yt-dlp
youtube-search
youtube-search-python
